backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
python-multipart==0.0.6
langchain
langchain-openai
langgraph
tensorflow
tensorflow_hub
numpy
Pillow
python-dotenv
